requirements.txt
biopython
pandas
scikit-learn
rich
